Add extra completion locations to the Camp map
The top right corner of the Camp map now contains the Collective
Unconscious. The state of all level completions can be seen there.

This puts a level completion location in both the relevant map, so that
the logic for a single level is all within one map, and the Collective
Unconscious because that is where the location is actually found.

If automatic map switching is added in the future, the Collective
Unconscious appears to be in the same level as Sasha's Lab, so they must
be part of the same map and, currently, Sasha's Lab is in the Camp map.