Revert MMI2 logic back to the current (incorrect) logic
Shield should be required, however it is currently missing from the
logic.

So while the logic is wrong in general, this fixes the tracker being
different to the currently used logic.

Technically, a Dream Fluff can be used to survive the sniper shot, but
that's not something that would likely be considered 'normal' logic (and
additionally requires access to the camp store).