Adjust "(MM Neighborhood) Post Office Lobby" to match current logic
Currently, this location is in the wrong region because it does not
require the Cobweb Duster, however, the tracker had pre-emptively moved
it to the correct region.

The logic for this location is now set to optionally require the Cobweb
Duster, showing that it can be acquired, but is technically out of
logic.

The region for this location will likely be adjusted in a future update
to the apworld and then this commit can be reverted.