Add settings tab for yaml settings and Goal locations
Goal locations are hidden when that Goal is not applicable with the
current settings. This requires PopTracker 0.26+, so that is the new
minimum PopTracker version.

To simplify the presentation of Goals to the user, the two yaml options
for Goals are split into three settings 'items', one for each Goal
requirement.

None of the Goal locations nor new settings are synced with AP because
the Goal locations do not appear to send events when checked and none of
the yaml options appear to be present in AP slot data at this time.

settings_items.json previously contained settings for the pack itself
and has been renamed to pack_settings_items.json.