Automatic addition/removal of Levitation when toggling StartingLevitation
Implemented using an item code watch to watch for when
StartingLevitation is toggled and then add/remove a level of Levitation.

I had to add a hidden item to track state, otherwise reloading the pack
would continuously add extra levitation items each time the pack was
reloaded with StartingLevitation enabled in the autosaved tracker state.