Out-of-logic (BB Area 2) Trapeze Cobweb without Cobweb Duster
Using Marksmanship Raz's head can pass through walls and iteract with
items on the other side. The item is too far away from the Cobweb for
Raz to grab it by using Marksmanship against the Cobweb (unless it's
Baggage).

However, the item is close to the back wall, so Raz can use Marksmanship
against this wall instead. The problem is that there is no ground to
stand on by the back wall, so Raz must use Marksmanship while falling
through the air in order to pass through the back wall and interact with
the item. Fortunately, there is at least ground at the bottom and some
ropes in the air, but this is still a difficult trick to pull off.

Having Levitation makes the trick a bit easier, since Raz can float to
about the right height, release Levitation to start falling and then
immediately use Marksmanship.