Out-of-logic (WW Main) Blacksmith's Right Building with levitation
The item is placed close enough to the wall that Raz collides with the
item from the other side of the wall, so perform one of the out-of-logic
tricks to get into Blacksmith's Left Building, but deliberately go to
the roof and jump across to the roof of the right Building. From there,
jump over the roof and float down the back wall of the right building
along where the item is on the other side.